Move controllers and concerns from shared-web to main application
