+++ cosmetics-web/app/controllers/concerns/cache_concern.rb
@@ -1,0 +1,9 @@
+module CacheConcern
+  extend ActiveSupport::Concern
+
+  def set_cache_headers
+    response.headers["Cache-Control"] = "no-cache, no-store"
+    response.headers["Pragma"] = "no-cache"
+    response.headers["Expires"] = "Fri, 01 Jan 1990 00:00:00 GMT"
+  end
+end
