--- shared-web/app/controllers/shared/web/concerns/cache_concern.rb
@@ -1,15 +1,0 @@
-module Shared
-  module Web
-    module Concerns
-      module CacheConcern
-        extend ActiveSupport::Concern
-
-        def set_cache_headers
-          response.headers["Cache-Control"] = "no-cache, no-store"
-          response.headers["Pragma"] = "no-cache"
-          response.headers["Expires"] = "Fri, 01 Jan 1990 00:00:00 GMT"
-        end
-      end
-    end
-  end
-end
