--- shared-web/app/controllers/shared/web/concerns/http_auth_concern.rb
@@ -1,18 +1,0 @@
-module Shared
-  module Web
-    module Concerns
-      module HttpAuthConcern
-        extend ActiveSupport::Concern
-
-        included do
-          if ENV["BASIC_AUTH_USERNAME"].present? && ENV["BASIC_AUTH_PASSWORD"].present?
-            http_basic_authenticate_with(
-              name: ENV.fetch("BASIC_AUTH_USERNAME"),
-              password: ENV.fetch("BASIC_AUTH_PASSWORD")
-            )
-          end
-        end
-      end
-    end
-  end
-end
